auth0-ai-samples/authorization-for-rag/langchain-python/requirements.txt
python-dotenv==1.0.1
langchain==0.3.14
langchain-community==0.3.14
openfga-sdk==0.9.0
faiss-cpu==1.9.0.post1
langchain-openai==0.3.0
